fix: rename app/email.py to app/mailer.py + apply ruff format
app/email.py shadowed Python's stdlib 'email' package. SQLAlchemy's
util.compat does 'from importlib import metadata', which internally
imports 'email' — and got our app/email.py instead of stdlib, dying
with 'No module named email.utils; email is not a package'. This
broke test-backend (prestart could not import sqlalchemy) and
test-docker-compose (backend container failed to start).

Also includes the ruff-format diff that lint-backend was flagging.
