+++ backend/app/models/showtime_visibility.py
@@ -22,6 +22,7 @@ from app.utils import now_amsterdam_naive
 
 class ShowtimeVisibilitySetting(SQLModel, table=True):
     """Whether a user's attendance for a showtime is visible to all friends or restricted."""
+
     owner_id: UUID = Field(
         foreign_key="user.id",
         primary_key=True,
@@ -38,6 +39,7 @@ class ShowtimeVisibilitySetting(SQLModel, table=True):
 
 class ShowtimeVisibilityFriend(SQLModel, table=True):
     """Explicit friend-level allow: viewer_id may see owner's attendance for showtime_id."""
+
     owner_id: UUID = Field(
         foreign_key="user.id",
         primary_key=True,
@@ -58,6 +60,7 @@ class ShowtimeVisibilityFriend(SQLModel, table=True):
 
 class ShowtimeVisibilityGroup(SQLModel, table=True):
     """Group-level allow: all members of group_id may see owner's attendance for showtime_id."""
+
     owner_id: UUID = Field(
         foreign_key="user.id",
         primary_key=True,
